Refactor: Split TouchCarousel.tsx into smaller components
Refactored `TouchCarousel.tsx` into smaller, more manageable components to improve code organization and maintainability. No functionality changes were introduced. Unused imports and files were removed.
