+++ src/components/dashboard/carousel/useTouchCarousel.ts
@@ -1,0 +1,81 @@
+
+import { useCallback, useEffect, useState } from 'react';
+import { EmblaOptionsType } from 'embla-carousel';
+import useEmblaCarousel from 'embla-carousel-react';
+import { useIsMobile } from '@/hooks/use-mobile';
+
+interface UseTouchCarouselProps {
+  coursesLength: number;
+}
+
+export const useTouchCarousel = ({ coursesLength }: UseTouchCarouselProps) => {
+  const isMobile = useIsMobile();
+  const [canScrollPrev, setCanScrollPrev] = useState(false);
+  const [canScrollNext, setCanScrollNext] = useState(false);
+  const [selectedIndex, setSelectedIndex] = useState(0);
+
+  // Optimized Embla configuration for mobile touch with snap
+  const options: EmblaOptionsType = {
+    align: 'center',
+    slidesToScroll: 1,
+    containScroll: 'trimSnaps',
+    dragFree: false, // Disable drag free for snap behavior
+    skipSnaps: false,
+    startIndex: 0,
+    // Add snap behavior
+    loop: false,
+  };
+
+  const [emblaRef, emblaApi] = useEmblaCarousel(options);
+
+  const scrollPrev = useCallback(() => {
+    if (emblaApi) {
+      console.log('TouchCarousel: Scrolling to previous slide');
+      emblaApi.scrollPrev();
+    }
+  }, [emblaApi]);
+
+  const scrollNext = useCallback(() => {
+    if (emblaApi) {
+      console.log('TouchCarousel: Scrolling to next slide');
+      emblaApi.scrollNext();
+    }
+  }, [emblaApi]);
+
+  const onSelect = useCallback(() => {
+    if (!emblaApi) return;
+    setCanScrollPrev(emblaApi.canScrollPrev());
+    setCanScrollNext(emblaApi.canScrollNext());
+    setSelectedIndex(emblaApi.selectedScrollSnap());
+    console.log('TouchCarousel: Carousel state updated', {
+      canScrollPrev: emblaApi.canScrollPrev(),
+      canScrollNext: emblaApi.canScrollNext(),
+      selectedIndex: emblaApi.selectedScrollSnap(),
+      totalSlides: emblaApi.scrollSnapList().length
+    });
+  }, [emblaApi]);
+
+  useEffect(() => {
+    if (!emblaApi) return;
+    console.log('TouchCarousel: Embla API initialized with', coursesLength, 'courses');
+    onSelect();
+    emblaApi.on('select', onSelect);
+    emblaApi.on('reInit', onSelect);
+
+    return () => {
+      emblaApi.off('select', onSelect);
+      emblaApi.off('reInit', onSelect);
+    };
+  }, [emblaApi, onSelect, coursesLength]);
+
+  return {
+    emblaRef,
+    emblaApi,
+    canScrollPrev,
+    canScrollNext,
+    selectedIndex,
+    scrollPrev,
+    scrollNext,
+    isMobile
+  };
+};
